Modified the ux of the story_gen page. started working on the weather page
There were some flaws in the front-end of story gen page such as utilization of fixed pixel heights instead of dynamic causing parts of the page to be not shown while in minimized form. this part has been fixed.

A new feature has been added such that when you press on the buttons to generate the story a side box will appear where the story will be shown.

UI design has to be improve for the story gen page as we are using default values.

fronted on the weather page has been started.